requirements.txt
pydantic==2.4.2
pyjwt-2.8.0
